+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Print some counts when running `zola check`
 - Re-render all pages/sections when `anchor-link.html` is changed
 - Taxonomies can now have the same name in multiple languages
+- `zola init` can now be create sites inside the current directory
 
 ## 0.8.0 (2019-06-22)